Update font_profont_10_72dpi.h
Replaced the glyph for !  It was blank - never displayed

The fonts in font_profont_10_72dpi.h are messed up for the ! and ' and | bitmaps (or glyphs). I can see why they're printing what they are. For instance, the font bitmap for the | char is:
const uint8_t profont_10_72_0x7c[] __attribute__((__progmem__)) = { /* '|' width: 6 */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0xc0, /* [*] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
}; 

The bitmap shown in the comments shows the single pixel in the middle of the glyph, and the setting of 0x00s in the rows with the single 0xc0 in a middle row matches that.

Same for the !, it's all 0x00s:
const uint8_t profont_10_72_0x21[] __attribute__((__progmem__)) = { /* '!' width: 6 */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
};
and the ', all 0x00s:
const uint8_t profont_10_72_0x27[] __attribute__((__progmem__)) = { /* ''' width: 6 */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
0x00, /* [ ] */ 
};
That's why they're blank.

Now, I'm not sure what they *should* be, since I can't quite decode the bitmap values (not sure what the . and * are in the comments for the glyph 'rows'), but looking at other chars gives a hint. For instance, the [ char is close to the |. Since the [ char is:
const uint8_t profont_10_72_0x5b[] __attribute__((__progmem__)) = { /* '[' width: 6 */ 
0xf0, /* [**] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xf0, /* [**] */ 
};
then that tells me we could get | by changing the top & bottom 0xf0 to 0xc0.

So, change the corresponding arrays in the file font_profont_10_72dpi.h with these 3:
const uint8_t profont_10_72_0x7c[] __attribute__((__progmem__)) = { /* '|' width: 6 */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 

};

A ' is close to a |, maybe just the top 3 rows, so it could be:
const uint8_t profont_10_72_0x27[] __attribute__((__progmem__)) = { /* ''' width: 6 */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
};

And a ! is close to |, but not as tall and with a blank row on line up from the bottom, and a little fatter up top:
const uint8_t profont_10_72_0x21[] __attribute__((__progmem__)) = { /* '!' width: 6 */ 
0x00, /* [ ] */ 
0xf0, /* [**] */ 
0xf0, /* [**] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0xc0, /* [* ] */ 
0x00, /* [ ] */ 
0xc0, /* [* ] */ 
0x00, /* [ ] */
};